// SP.Web/models/SelectedCourse.ts
export class SelectedCourse {
  enrollmentOption: string;
  course: string;
  courseCode: string;
  credit: number;
  specialization: string;
  category: string;
  level: string;
  scoring: boolean = false;
  assignedTo: string;
  department: string;
  program: string;
  specializationName: string;
}